lcs-sacoche_trunk : update to version 2015_09_02
git-svn-id: https://svn.tice.ac-caen.fr/svn/LcsSE3/LCS/LCS2.0ndg/sources/trunk@5410 1012e355-7652-49d3-8fd6-1f9b0b464cea

+++ lcs-sacoche/Sacoche/_sql/structure/sacoche_groupe.sql
@@ -5,7 +5,7 @@ CREATE TABLE sacoche_groupe (
   groupe_type    ENUM("classe","groupe","besoin","eval")                      COLLATE utf8_unicode_ci NOT NULL DEFAULT "classe",
   groupe_ref     CHAR(8)                                                      COLLATE utf8_unicode_ci NOT NULL DEFAULT "",
   groupe_nom     VARCHAR(20)                                                  COLLATE utf8_unicode_ci NOT NULL DEFAULT "",
-  niveau_id      SMALLINT(5)                                                  UNSIGNED                NOT NULL DEFAULT 0,
+  niveau_id      MEDIUMINT(8)                                                 UNSIGNED                NOT NULL DEFAULT 0,
   fiche_brevet   ENUM("","1vide","2rubrique","3mixte","4synthese","5complet") COLLATE utf8_unicode_ci NOT NULL DEFAULT "",
   PRIMARY KEY (groupe_id),
   KEY niveau_id (niveau_id),
