+++ lcs-sacoche/Sacoche/_sql/structure/sacoche_referentiel_domaine.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS sacoche_referentiel_domaine;
 CREATE TABLE sacoche_referentiel_domaine (
   domaine_id    SMALLINT(5)  UNSIGNED                NOT NULL AUTO_INCREMENT,
   matiere_id    SMALLINT(5)  UNSIGNED                NOT NULL DEFAULT 0,
-  niveau_id     SMALLINT(5)  UNSIGNED                NOT NULL DEFAULT 0,
+  niveau_id     MEDIUMINT(8) UNSIGNED                NOT NULL DEFAULT 0,
   domaine_ordre TINYINT(3)   UNSIGNED                NOT NULL DEFAULT 1 COMMENT "Commence à 1.",
   domaine_ref   CHAR(1)      COLLATE utf8_unicode_ci NOT NULL DEFAULT "",
   domaine_nom   VARCHAR(128) COLLATE utf8_unicode_ci NOT NULL DEFAULT "",
